refactor, moved config swagger to config folder
