+++ src/main/java/com/endava/endabank/configuration/SwaggerConfig.java
@@ -1,11 +1,8 @@
-package com.endava.endabank;
+package com.endava.endabank.configuration;
 
-
-import java.util.Collections;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import springfox.documentation.builders.PathSelectors;
@@ -14,10 +11,10 @@ import springfox.documentation.service.ApiInfo;
 import springfox.documentation.service.Contact;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+import java.util.Collections;
 
 @Configuration
-@EnableSwagger2
 public class SwaggerConfig implements WebMvcConfigurer {
 
     @Override
